remove config explorer, update tests
