+++ README.md
@@ -1,12 +1,10 @@
 # AllenNLP Config Explorer and Simple Server
 
-This repo contains a config explorer for interactively building AllenNLP configuration files, as well as a simple demo server for serving a model.  To file an issue, please open a ticket on https://github.com/allenai/allennlp and tag it with "Server".
+This repo contains a simple demo server for serving a model. To file an issue, please open a ticket on https://github.com/allenai/allennlp and tag it with "Server".
 
 ## Installation
 
-### Installing from source
-
-You can also install AllenNLP Server by cloning our git repository:
+You can install AllenNLP Server by cloning our git repository:
 
 ```bash
 git clone https://github.com/allenai/allennlp-server
@@ -20,15 +18,7 @@ pip install --editable .
 
 ## Running AllenNLP Server
 
-AllenNLP Server is a plugin for AllenNLP, so new subcommands will be added to it.
-
-To run the configuration explorer, see:
-
-```bash
-allennlp configure --help
-```
-
-To run the simple demo server, see:
+AllenNLP Server is a plugin for AllenNLP which adds a "serve" subcommand:
 
 ```bash
 allennlp serve --help
