+++ allennlp_server/commands/__init__.py
@@ -1,2 +1,1 @@
-from allennlp_server.commands.configure import Configure
 from allennlp_server.commands.server_simple import SimpleServer
