--- allennlp_server/config_explorer/__init__.py
@@ -1,1 +1,0 @@
-from allennlp_server.config_explorer.config_explorer import make_app
