+++ setup.cfg
@@ -45,7 +45,6 @@ python_requires = >=3.6.1
 
 [options.packages.find]
 include =
-  allennlp_plugins.*
   allennlp_server
 exclude =
   *.tests
